Move Sapling transaction components to a separate module.
